Not the text but one of the images
One of the images on the right is of an east asian woman. There are very few women who wear their hair like that now. It sticks out as an outdated caricature. Something to consider ;)
https://assets.exercism.io/assets/people/5-386a4c169c581ba5f04342359f2db34d7371b9853a7f453a85d8eadcdc0607b6.png

Other than that, I'm having a great time on Exercism so far. Thank you team!!